pyproject.toml
[tool.poetry]
name = "orionx-python-client"
version = "0.3.10"
description = ""
authors = ["adolfrodeno <amvillalobos@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.1"
pydantic = "^1.8.2"
ujson = "^5.1.0"
requests = ">=2.25.0"
aiohttp = "^3.8.1"
aiodns = "^3.0.0"
cchardet = "^2.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
